requirements.txt
Django==2.0.7
psycopg2==2.7.5
pytz==2018.5
django-extensions==2.1.2
django-filter==2.0.0
djangorestframework==3.8.2
ipython==6.5.0
Markdown==2.1.0